requirements.txt
# === 高级监控依赖 ===
prometheus-client>=0.19.0       # Prometheus监控客户端
aiohttp>=3.8.0                  # 异步HTTP客户端（用于Webhook告警）

# Web框架和API
fastapi==0.111.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据库相关
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.2

# 环境配置
python-dotenv==1.0.1

# 认证和安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# 数据处理和科学计算
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0

# 图片和文档处理
Pillow==10.3.0

# AI和机器学习
sentence-transformers==2.5.1
transformers==4.38.2
torch==2.2.1
openai==1.12.0

# 文档处理
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.5
xlrd==2.0.1
python-pptx==0.6.23

# 文本转语音
gTTS==2.5.1

# HTTP请求
requests==2.32.3
httpx==0.27.2
aiohttp>=3.8.0  # 异步HTTP客户端，用于LLM性能测试

# 数据验证
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.1.2

# 类型检查支持
typing-extensions>=4.0.0  # LLM路由类型安全增强

# 工具库
python-dateutil==2.8.2
pytz==2024.1

# 异步支持
asyncio-mqtt==0.16.1

# 缓存和性能优化
redis==5.0.1
python-magic>=0.4.27; platform_system == "Windows"  # 仅在Windows安装，修复语法警告
bleach==6.1.0

# 知识库增强功能依赖
jieba==0.42.1                   # 中文分词，用于智能推荐
schedule==1.2.0                 # 任务调度，用于监控告警

# 开发和调试工具（可选）
pytest>=8.0.0  # 更新版本以支持LLM增强功能测试
pytest-asyncio>=0.23.0  # 更新版本以支持异步测试

boto3==1.37.38

# === LLM路由增强功能依赖 ===
# 以上已集成的依赖：
# - aiohttp>=3.8.0: 异步HTTP客户端，用于性能测试和监控
# - pytest>=8.0.0: 测试框架，支持LLM功能单元测试
# - pytest-asyncio>=0.23.0: 异步测试支持
# - typing-extensions>=4.0.0: 类型检查支持，提升代码安全性
# 注意：statistics 是Python标准库，无需额外安装

# 安全增强依赖（已安装）
filetype==1.2.0  # 文件类型检测增强
plyara==2.1.1     # YARA规则解析器

# 高级安全依赖（需要编译器，Windows上较复杂）
# yara-python==4.3.1  # 恶意软件检测引擎
# 安装说明：需要先安装 Visual Studio Build Tools
# 下载地址：https://visualstudio.microsoft.com/visual-cpp-build-tools/
